Make paginated collection trait
